% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hewlett-packard.r
\docType{data}
\name{Agilent_leds}
\alias{Agilent_leds}
\title{Spectral data for LEDs supplied by Agilent/Hewlett Packard}
\format{
A vector of character strings.
}
\usage{
Agilent_leds
}
\description{
The collection of spectra \code{\link{leds.mspct}} contains spectra for 
some light emitting diodes (LEDs) from Agilent/Hewlett Packard (ca. 1995). 
Data are normalized to one at the wavelength of maximum emission. The
character vector \code{Agilent_leds} contains the names of the spectra to
facilitate their extraction from the collection.
}
\note{
The division of Hewlett Packard which supplied these LEDs became part of
Agilent when this division spin-off the mother company. More recently
the electronic components division of Agilent became Avago Technologies for
a while. Currently, BROADCOM supplies some of these LEDs or similar improved 
types.
}
\examples{
Agilent_leds

}
\references{
\url{https://www.broadcom.com/products/leds-and-displays/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{HueyJann_leds}},
\code{\link{LedEngin_leds}},
\code{\link{Ledguhon_leds}},
\code{\link{Marktech_leds}},
\code{\link{Nichia_leds}},
\code{\link{Norlux_leds}},
\code{\link{Osram_leds}},
\code{\link{QuantumDevices_leds}},
\code{\link{Roithner_leds}},
\code{\link{TaoYuan_leds}},
\code{\link{Weili_leds}}
}
\concept{manufacturers}
\keyword{datasets}
