% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huey-jann.r
\docType{data}
\name{huey_jann}
\alias{huey_jann}
\title{Spectral data for LEDs array supplied by Huey Jann}
\format{A vector of character strings.}
\usage{
huey_jann
}
\description{
Datasets containing the wavelengths and
tabulated values spectral emittance for different light emitting diodes (LEDs) arrays
from Huey Jann Electronics Industry Co., Ltd. (Taiwan).
Absolute values are not meaningful as the measuring distances are variable, and in most cases unknown.
}
\details{
The variables are as follows:
\itemize{
  \item w.length (nm)  
  \item s.e.irrad (W m-2 nm-1)
}
}
\note{
Huey Jann was a Taiwanese supplier of LED array. It is no longer in
business.
}
\examples{
huey_jann
leds.mspct[huey_jann]

}
\keyword{datasets}

