% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huey-jann.r
\docType{data}
\name{HueyJann_leds}
\alias{HueyJann_leds}
\title{Spectral data for LEDs arrays supplied by Huey Jann}
\format{
A vector of character strings.
}
\usage{
HueyJann_leds
}
\description{
The collection of spectra \code{\link{leds.mspct}} contains spectra for some
arrays of light emitting diodes (LEDs) from Huey Jann Electronics Industry
Co., Ltd. (Taiwan). Data are normalized to one at the wavelength of maximum
emission. The character vector \code{HueyJann_leds} contains the names of the
spectra to facilitate their extraction from the collection.
}
\note{
Huey Jann was a Taiwanese supplier of LED array. It is no longer in
business.
}
\examples{
HueyJann_leds

}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{Agilent_leds}},
\code{\link{LedEngin_leds}},
\code{\link{Ledguhon_leds}},
\code{\link{Marktech_leds}},
\code{\link{Nichia_leds}},
\code{\link{Norlux_leds}},
\code{\link{Osram_leds}},
\code{\link{QuantumDevices_leds}},
\code{\link{Roithner_leds}},
\code{\link{TaoYuan_leds}},
\code{\link{Weili_leds}}
}
\concept{manufacturers}
\keyword{datasets}
