% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/led-engin.r
\docType{data}
\name{led_engin}
\alias{led_engin}
\title{Spectral data for LEDs array supplied by Led Engin}
\format{A vector of character strings.}
\usage{
led_engin
}
\description{
Datasets containing the wavelengths and tabulated values spectral emittance 
for light emitting diodes (LEDs) and arrays from Led Engin (USA). Data are
normalized to one at the wavelength of maximum emission.
}
\details{
The variables are as follows:
\itemize{
  \item w.length (nm)  
  \item s.e.irrad (W m-2 nm-1)
}
}
\note{
Led Engin is a supplier of power LEDs of high efficiency.
\url{http://www.ledengin.com/}
}
\examples{
led_engin
# leds.mspct[led_engin]

}
\keyword{datasets}

