% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumitronix.r
\docType{data}
\name{lumitronix}
\alias{lumitronix}
\title{Spectral data for LED array from LUMITRONIX}
\format{A vector of character strings.}
\usage{
lumitronix
}
\description{
Dataset containing the wavelengths and tabulated values spectral emittance 
for a high power light emitting diode (LED) array from LUMITRONIX  based 
NICHIA's high efficiency natural white SMT LEDs. Specifications: LUMITRONIX 
SmartArray Q36 LED-Module, 4247 lm, 4000K, 39W electrical. Data are
normalized to one at the wavelength of maximum emission.
}
\details{
The variables are as follows: 
\itemize{
\item w.length (nm) 
\item s.e.irrad  (W m-2 nm-1) 
}
}
\note{
Lumitronix is a supplier of LED arrays, and a distributor of LEDs.
\url{http://www.leds.de/}
}
\examples{
lumitronix
leds.mspct[lumitronix]

}
\keyword{datasets}

