% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marktech.r
\docType{data}
\name{marktech}
\alias{marktech}
\title{Spectral data for LEDs array supplied by Marktech}
\format{
A vector of character strings.
}
\usage{
marktech
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for light emitting diodes (LEDs) from Marktech. 
Data are normalized to one at the wavelength of maximum emission.
}
\examples{
marktech

}
\references{
\url{http://www.marktechopto.com/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{hewlett_packard}},
\code{\link{huey_jann}},
\code{\link{led_engin}},
\code{\link{leds_global}},
\code{\link{lumitronix}},
\code{\link{nichia}},
\code{\link{norlux}},
\code{\link{osram}},
\code{\link{quantum_devices}},
\code{\link{roithner_laser}},
\code{\link{seti}},
\code{\link{tao_yuan}},
\code{\link{unknown}}
}
\concept{manufacturers}
\keyword{datasets}
