% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichia.r
\docType{data}
\name{nichia}
\alias{nichia}
\title{Spectral data for LEDs array supplied by NICHIA}
\format{
A vector of character strings.
}
\usage{
nichia
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for light emitting diodes (LEDs) from NICHIA. 
Data are normalized to one at the wavelength of maximum emission.
}
\examples{
nichia

}
\references{
\url{http://www.nichia.com/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{hewlett_packard}},
\code{\link{huey_jann}},
\code{\link{led_engin}},
\code{\link{leds_global}},
\code{\link{lumitronix}},
\code{\link{marktech}},
\code{\link{norlux}},
\code{\link{osram}},
\code{\link{quantum_devices}},
\code{\link{roithner_laser}},
\code{\link{seti}},
\code{\link{tao_yuan}},
\code{\link{unknown}}
}
\concept{manufacturers}
\keyword{datasets}
