% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichia.r
\docType{data}
\name{Nichia_leds}
\alias{Nichia_leds}
\title{Spectral data for LEDs from NICHIA}
\format{
A vector of character strings.
}
\usage{
Nichia_leds
}
\description{
The collection of spectra \code{\link{leds.mspct}} contains spectra for 
some light emitting diodes (LEDs) from Nichia, Japan. 
Data are normalized to one at the wavelength of maximum emission. The
character vector \code{Nichia_leds} contains the names of the spectra to
facilitate their extraction from the collection.
}
\note{
Some of these LEDs were supplied by LUMITRONIX, Germany, assembled in
arrays of the series names SmartArray and LinearZ.
}
\examples{
Nichia_leds

}
\references{
\url{https://www.nichia.com/}, \url{https://www.leds.de/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{Agilent_leds}},
\code{\link{HueyJann_leds}},
\code{\link{LedEngin_leds}},
\code{\link{Ledguhon_leds}},
\code{\link{Marktech_leds}},
\code{\link{Norlux_leds}},
\code{\link{Osram_leds}},
\code{\link{QuantumDevices_leds}},
\code{\link{Roithner_leds}},
\code{\link{TaoYuan_leds}},
\code{\link{Weili_leds}}
}
\concept{manufacturers}
\keyword{datasets}
