% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norlux.r
\docType{data}
\name{norlux}
\alias{norlux}
\title{Spectral data for LEDs array supplied by Norlux}
\format{A vector of character strings.}
\usage{
norlux
}
\description{
Datasets containing the wavelengths and tabulated values spectral emittance 
for the NHXRGB0905005 light emitting diodes (LEDs) arrays from Norlux (USA). 
Data are normalized to one at the wavelength of maximum emission.
}
\details{
The variables are as follows:
\itemize{
  \item w.length (nm)  
  \item s.e.irrad (W m-2 nm-1)
}
}
\note{
Norlux is now part of Thomas Research Products.
\url{http://www.trpssl.com/}
}
\examples{
norlux
leds.mspct[norlux]

}
\keyword{datasets}

