% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norlux.r
\docType{data}
\name{Norlux_leds}
\alias{Norlux_leds}
\title{Spectral data for LED COB arrays supplied by Norlux}
\format{
A vector of character strings.
}
\usage{
Norlux_leds
}
\description{
The collection of spectra \code{\link{leds.mspct}} contains spectra for 
some light emitting diodes (LEDs) from Norlux, U.S.A. 
Data are normalized to one at the wavelength of maximum emission. The
character vector \code{Norlux_leds} contains the names of the spectra to
facilitate their extraction from the collection.
}
\note{
Norlux is now part of Thomas Research Products.
}
\examples{
Norlux_leds

}
\references{
\url{https://www.trpssl.com/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{Agilent_leds}},
\code{\link{HueyJann_leds}},
\code{\link{LedEngin_leds}},
\code{\link{Ledguhon_leds}},
\code{\link{Marktech_leds}},
\code{\link{Nichia_leds}},
\code{\link{Osram_leds}},
\code{\link{QuantumDevices_leds}},
\code{\link{Roithner_leds}},
\code{\link{TaoYuan_leds}},
\code{\link{Weili_leds}}
}
\concept{manufacturers}
\keyword{datasets}
