% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-maya-leds.r
\docType{data}
\name{oo_maya_leds}
\alias{oo_maya_leds}
\title{Spectra acquired with Ocean Optics Maya2000 Pro}
\format{
A vector of character strings.
}
\usage{
oo_maya_leds
}
\description{
Dataset containing the wavelengths and tabulated values spectral
  emittance for LEDs. Data for both low power single-die devices and high
  power LEDs arrays are included. The data are normalized to the peak of
  maximum spectral energy irradiance.
}
\details{
\code{oo_maya_leds} is a character vector with indexes to members of
  \code{\link{leds.mspct}}.

The variables in each member spectrum are as follows: \itemize{ \item
w.length (nm) \item s.e.irrad }
}
\note{
Instrument used: Ocean Optics Maya2000 Pro single-monochromator array
  spectroradiometer with a Bentham cosine corrected input optics. A complex
  set of corrections and calibration procedure used. Raw spectral data
  processed with R packages 'MayaCalc' or 'ooacquire'.
}
\examples{
oo_maya_leds

}
\seealso{
\code{\link{leds.mspct}}
}
\keyword{datasets}
\keyword{internal}
