% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osram.r
\docType{data}
\name{osram}
\alias{osram}
\title{Spectral data for LEDs array supplied by Osram}
\format{A vector of character strings.}
\usage{
osram
}
\description{
Datasets containing the wavelengths and tabulated values spectral emittance 
for the NHXRGB0905005 light emitting diodes (LEDs) arrays from Osram. 
Data are normalized to one at the wavelength of maximum emission.
}
\details{
The variables are as follows:
\itemize{
  \item w.length (nm)  
  \item s.e.irrad (W m-2 nm-1)
}
}
\note{
Current trade name is Osram Opto Semiconductors
\url{http://www.osram-os.com/}
}
\examples{
osram
leds.mspct[osram]

}
\keyword{datasets}

