% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osram.r
\docType{data}
\name{Osram_leds}
\alias{Osram_leds}
\title{Spectral data for LEDs array supplied by Osram}
\format{
A vector of character strings.
}
\usage{
Osram_leds
}
\description{
The collection of spectra \code{\link{leds.mspct}} contains spectra for 
some light emitting diodes (LEDs) from Osram Opto Semiconductors. 
Data are normalized to one at the wavelength of maximum emission. The
character vector \code{Osram_leds} contains the names of the spectra to
facilitate their extraction from the collection.
}
\note{
Current trade name is Osram Opto Semiconductors
}
\examples{
Osram_leds

}
\references{
\url{https://www.osram-os.com/}
}
\seealso{
\code{\link{leds.mspct}} and \code{\link{LedEngin_leds}}

Other manufacturers: 
\code{\link{Agilent_leds}},
\code{\link{HueyJann_leds}},
\code{\link{LedEngin_leds}},
\code{\link{Ledguhon_leds}},
\code{\link{Marktech_leds}},
\code{\link{Nichia_leds}},
\code{\link{Norlux_leds}},
\code{\link{QuantumDevices_leds}},
\code{\link{Roithner_leds}},
\code{\link{TaoYuan_leds}},
\code{\link{Weili_leds}}
}
\concept{manufacturers}
\keyword{datasets}
