% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum-devices.r
\docType{data}
\name{quantum_devices}
\alias{quantum_devices}
\title{Spectral data for LEDs array supplied by Quantum Devices}
\format{A vector of character strings.}
\usage{
quantum_devices
}
\description{
Datasets containing the wavelengths and tabulated values spectral emittance 
for light emitting diodes (LEDs) from Quantum Devices (USA). Data are
normalized to one at the wavelength of maximum emission.
}
\details{
The variables are as follows:
\itemize{
  \item w.length (nm)  
  \item s.e.irrad (W m-2 nm-1)
}
}
\note{
Quantum Devices produces both individual LEDs and luminaires.
\url{http://www.quantumdev.com/}
}
\examples{
quantum_devices
leds.mspct[quantum_devices]

}
\keyword{datasets}

