% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roithner-laser.r
\docType{data}
\name{roithner_laser}
\alias{roithner_laser}
\title{Spectral data for LEDs supplied by Roithner Laser}
\format{A vector of character strings.}
\usage{
roithner_laser
}
\description{
Datasets containing the wavelengths and tabulated values for spectral
emittance for different light emitting diodes (LEDs) and LED arrays supplied
by Roithner Laser (Austria). Data are normalized to one at the wavelength of
maximum emission.
}
\details{
The variables are as follows:
\itemize{
  \item w.length (nm)
  \item s.e.irrad (W m-2 nm-1)
}
}
\note{
Roithner LaserTechnik is a distributor and reseller of LEDs, LED arrays and
lasers. They have a very extensive catalogue covering almost wavelengths for
which LEDs are manufactured.
\url{http://www.roithner-laser.com/}
}
\examples{
roithner_laser
leds.mspct[roithner_laser]

}
\keyword{datasets}

