% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roithner-laser.r
\docType{data}
\name{Roithner_leds}
\alias{Roithner_leds}
\title{Spectral data for LEDs supplied by Roithner Laser}
\format{
A vector of character strings.
}
\usage{
Roithner_leds
}
\description{
The collection of spectra \code{\link{leds.mspct}} contains spectra for 
some light emitting diodes (LEDs) from Roithner LaserTechnik, Austria. 
Data are normalized to one at the wavelength of maximum emission. The
character vector \code{Roithner_leds} contains the names of the spectra to
facilitate their extraction from the collection.
}
\details{
Names of datasets containing the wavelengths and tabulated values for
spectral emittance for different light emitting diodes (LEDs) and LED arrays
supplied by Roithner Laser (Austria). Data are normalized to one at the
wavelength of maximum emission.
}
\note{
Roithner LaserTechnik is a distributor and reseller of LEDs, LED arrays and
lasers. They have a very extensive catalogue covering almost all wavelengths
for which LEDs are manufactured. Many of the LEDs are sold under new codes
as they are retested and in some cases individual characterization data
provided. For example some of short UV LEDs sold are from SETi.
}
\examples{
Roithner_leds

}
\references{
\url{https://www.roithner-laser.com/} and 
\url{http://www.s-et.com/en/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{Agilent_leds}},
\code{\link{HueyJann_leds}},
\code{\link{LedEngin_leds}},
\code{\link{Ledguhon_leds}},
\code{\link{Marktech_leds}},
\code{\link{Nichia_leds}},
\code{\link{Norlux_leds}},
\code{\link{Osram_leds}},
\code{\link{QuantumDevices_leds}},
\code{\link{TaoYuan_leds}},
\code{\link{Weili_leds}}
}
\concept{manufacturers}
\keyword{datasets}
