% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seti.r
\docType{data}
\name{seti}
\alias{seti}
\title{Spectral data for LEDs array supplied by SETi}
\format{A vector of character strings.}
\usage{
seti
}
\description{
Datasets containing the wavelengths and tabulated values spectral emittance
for the NHXRGB0905005 light emitting diodes (LEDs) arrays from Osram.
Data are normalized to one at the wavelength of maximum emission.
}
\details{
The variables are as follows:
\itemize{
  \item w.length (nm)
  \item s.e.irrad (W m-2 nm-1)
}
}
\note{
SETi (Sensor Electronic Technologies) is a supplier of high power ultraviolet
LEDs emitting in the UVC, UVB and UVA regions of the spectrum. Many of these
LEDs are also sold under different type denominations by Roithner
LaserTechnik.
\url{http://www.s-et.com/}
}
\examples{
seti
leds.mspct[seti]

}
\keyword{datasets}

