% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tao-yuan.r
\docType{data}
\name{tao_yuan}
\alias{tao_yuan}
\title{Spectral data for LEDs supplied by TaoYuan}
\format{A vector of character strings.}
\usage{
tao_yuan
}
\description{
Datasets containing the wavelengths and tabulated values spectral emittance
for different light emitting diodes (LEDs) from TaoYuan Electron (HK). Data
are normalized to one at the wavelength of maximum emission.
}
\details{
The variables are as follows:
\itemize{
  \item w.length (nm)
  \item s.e.irrad (W m-2 nm-1)
}
}
\note{
TaoYuan Electron (HK) is a supplier of LEDs and LED arrays.
\url{http://www.ledwv.com/en/}
}
\examples{
tao_yuan
leds.mspct[tao_yuan]

}
\keyword{datasets}

