% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tao-yuan.r
\docType{data}
\name{TaoYuan_leds}
\alias{TaoYuan_leds}
\title{Spectral data for LEDs supplied by TaoYuan}
\format{
A vector of character strings.
}
\usage{
TaoYuan_leds
}
\description{
The collection of spectra \code{\link{leds.mspct}} contains spectra for 
some light emitting diodes (LEDs) from TaoYuan Electron (HK). 
Data are normalized to one at the wavelength of maximum emission. The
character vector \code{TaoYuan_leds} contains the names of the spectra to
facilitate their extraction from the collection.
}
\note{
TaoYuan Electron (HK) is a supplier of LEDs and LED arrays.
}
\examples{
TaoYuan_leds

}
\references{
\url{https://www.ledwv.com/en/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{Agilent_leds}},
\code{\link{HueyJann_leds}},
\code{\link{LedEngin_leds}},
\code{\link{Ledguhon_leds}},
\code{\link{Marktech_leds}},
\code{\link{Nichia_leds}},
\code{\link{Norlux_leds}},
\code{\link{Osram_leds}},
\code{\link{QuantumDevices_leds}},
\code{\link{Roithner_leds}},
\code{\link{Weili_leds}}
}
\concept{manufacturers}
\keyword{datasets}
