% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.r
\docType{data}
\name{unknown}
\alias{unknown}
\title{Spectral data for LEDs array of unknown manufacturer}
\format{A vector of character strings.}
\usage{
unknown
}
\description{
Datasets containing the wavelengths and tabulated values spectral emittance
for different "generic" light emitting diodes (LEDs) without type
specifications. Bought from shops like Class Ohlson or hobby targete
electronic suppliers. Absolute values are not meaningful as the measuring
distances are variable, and in most cases unknown.
}
\details{
The variables are as follows:
\itemize{
  \item w.length (nm)
  \item s.e.irrad (W m-2 nm-1)
}
}
\examples{
unknown
leds.mspct[unknown]

}
\keyword{datasets}

