% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shenzhen-weili.r
\docType{data}
\name{Weili_leds}
\alias{Weili_leds}
\alias{shenzhen_weili}
\title{Spectral data for LEDs array supplied by Shenzhen Weili Optical}
\format{
A vector of character strings.
}
\usage{
Weili_leds
}
\description{
The collection of spectra \code{\link{leds.mspct}} contains spectra for 
some light emitting diodes (LEDs) from Shenzhen Weili Optical. 
Data are normalized to one at the wavelength of maximum emission. The
character vector \code{Weili_leds} contains the names of the spectra to
facilitate their extraction from the collection.
}
\details{
Names of datasets containing the wavelengths and tabulated values spectral emittance
for different light emitting diodes (LEDs) arrays from Shenzhen Weili Optical
Ltd.  Data are normalized to one at the wavelength of maximum emission.
}
\note{
Leds Global and Shenzhen Weili are trade names of the same supplier of LED
arrays. They sell both standard types and also assemble customized arrays
upon request. Customized arrays may have up to twelve independent channels.
}
\examples{
Weili_leds

}
\references{
\url{https://www.leds-global.com/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{Agilent_leds}},
\code{\link{HueyJann_leds}},
\code{\link{LedEngin_leds}},
\code{\link{Ledguhon_leds}},
\code{\link{Marktech_leds}},
\code{\link{Nichia_leds}},
\code{\link{Norlux_leds}},
\code{\link{Osram_leds}},
\code{\link{QuantumDevices_leds}},
\code{\link{Roithner_leds}},
\code{\link{TaoYuan_leds}}
}
\concept{manufacturers}
\keyword{datasets}
