% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hewlett-packard.r
\docType{data}
\name{hewlett_packard}
\alias{hewlett_packard}
\title{Spectral data for LEDs array supplied by Agilent/Hewlett Packard}
\format{
A vector of character strings.
}
\usage{
hewlett_packard
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for light emitting diodes (LEDs) from Agilent/Hewlett Packard. 
Data are normalized to one at the wavelength of maximum emission.
}
\note{
The division of Hewlett Packard which supplied these LEDs became part of
Agilent when this division spin-off the mother company. More recently
the electronic components division of Agilent became Avago Technologies
which still supplies some of these LEDs or similar improved types.
}
\examples{
hewlett_packard

}
\references{
\url{https://www.broadcom.com/products/leds-and-displays/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{huey_jann}},
\code{\link{led_engin}},
\code{\link{leds_global}},
\code{\link{lumitronix}},
\code{\link{marktech}},
\code{\link{nichia}},
\code{\link{norlux}},
\code{\link{osram}},
\code{\link{quantum_devices}},
\code{\link{roithner_laser}},
\code{\link{seti}},
\code{\link{tao_yuan}},
\code{\link{unknown}}
}
\concept{manufacturers}
\keyword{datasets}
