% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huey-jann.r
\docType{data}
\name{huey_jann}
\alias{huey_jann}
\title{Spectral data for LEDs array supplied by Huey Jann}
\format{
A vector of character strings.
}
\usage{
huey_jann
}
\description{
Names of datasets containing the wavelengths and
tabulated values spectral emittance for different light emitting diodes (LEDs) arrays
from Huey Jann Electronics Industry Co., Ltd. (Taiwan).
Absolute values are not meaningful as the measuring distances are variable, and in most cases unknown.
}
\note{
Huey Jann was a Taiwanese supplier of LED array. It is no longer in
business.
}
\examples{
huey_jann

}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{hewlett_packard}},
\code{\link{led_engin}},
\code{\link{leds_global}},
\code{\link{lumitronix}},
\code{\link{marktech}},
\code{\link{nichia}},
\code{\link{norlux}},
\code{\link{osram}},
\code{\link{quantum_devices}},
\code{\link{roithner_laser}},
\code{\link{seti}},
\code{\link{tao_yuan}},
\code{\link{unknown}}
}
\concept{manufacturers}
\keyword{datasets}
