% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/led-engin.r
\docType{data}
\name{led_engin}
\alias{led_engin}
\title{Spectral data for LEDs array supplied by Led Engin}
\format{
A vector of character strings.
}
\usage{
led_engin
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for light emitting diodes (LEDs) and arrays from Led Engin (USA). Data are
normalized to one at the wavelength of maximum emission.
}
\note{
Led Engin is a supplier of power LEDs of high efficiency.
}
\examples{
led_engin

}
\references{
\url{http://www.ledengin.com/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{hewlett_packard}},
\code{\link{huey_jann}},
\code{\link{leds_global}},
\code{\link{lumitronix}},
\code{\link{marktech}},
\code{\link{nichia}},
\code{\link{norlux}},
\code{\link{osram}},
\code{\link{quantum_devices}},
\code{\link{roithner_laser}},
\code{\link{seti}},
\code{\link{tao_yuan}},
\code{\link{unknown}}
}
\concept{manufacturers}
\keyword{datasets}
