% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leds-mspct.r
\docType{data}
\name{leds.mspct}
\alias{leds.mspct}
\title{Spectral irradiance for diverse LEDs}
\format{A \code{"source_mspct"} object containing several
  \code{"source_spct"}.}
\usage{
leds.mspct
}
\description{
A collection of emission spectra of light-emitting-diodes
  from different suppliers.
}
\details{
The \code{"source_mspct"} object contains \code{"source_spct"}
  objects with spectral emission data.

  The variables in each member spectrum are as follows: \itemize{ \item
  w.length (nm) \item s.e.irrad (relative energy based units) }
}
\note{
Please see the help corresponding to each supplier for details.
}
\examples{
names(leds.mspct)
leds.mspct$UV395

}
\seealso{
\code{\link{oo_maya_leds}}
}
\keyword{datasets}

