% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shenzhen-weili.r
\docType{data}
\name{leds_global}
\alias{leds_global}
\alias{shenzhen_weili}
\title{Spectral data for LEDs array supplied by Shenzhen Weili Optical}
\format{
A vector of character strings.
}
\usage{
leds_global
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance
for different light emitting diodes (LEDs) arrays from Shenzhen Weili Optical
Ltd.  Data are normalized to one at the wavelength of maximum emission.
}
\note{
Leds Global and Shenzhen Weili are trade names of the same supplier of LED
arrays. They sell both standard types and also assemble customized arrays
upon request. Customized arrays may have up to five independent channels.
}
\examples{
shenzhen_weili
leds_global

}
\references{
\url{http://www.leds-global.com/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{hewlett_packard}},
\code{\link{huey_jann}},
\code{\link{led_engin}},
\code{\link{lumitronix}},
\code{\link{marktech}},
\code{\link{nichia}},
\code{\link{norlux}},
\code{\link{osram}},
\code{\link{quantum_devices}},
\code{\link{roithner_laser}},
\code{\link{seti}},
\code{\link{tao_yuan}},
\code{\link{unknown}}
}
\concept{manufacturers}
\keyword{datasets}
