% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumitronix.r
\docType{data}
\name{lumitronix}
\alias{lumitronix}
\title{Spectral data for LED array from LUMITRONIX}
\format{
A vector of character strings.
}
\usage{
lumitronix
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for a high power light emitting diode (LED) array from LUMITRONIX  based 
NICHIA's high efficiency natural white SMT LEDs. Specifications: LUMITRONIX 
SmartArray Q36 LED-Module, 4247 lm, 4000K, 39W electrical. Data are
normalized to one at the wavelength of maximum emission.
}
\note{
Lumitronix is a supplier of LED arrays, and a distributor of LEDs.
}
\examples{
lumitronix

}
\references{
\url{https://www.leds.de/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{hewlett_packard}},
\code{\link{huey_jann}},
\code{\link{led_engin}},
\code{\link{leds_global}},
\code{\link{marktech}},
\code{\link{nichia}},
\code{\link{norlux}},
\code{\link{osram}},
\code{\link{quantum_devices}},
\code{\link{roithner_laser}},
\code{\link{seti}},
\code{\link{tao_yuan}},
\code{\link{unknown}}
}
\concept{manufacturers}
\keyword{datasets}
