% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norlux.r
\docType{data}
\name{norlux}
\alias{norlux}
\title{Spectral data for LEDs array supplied by Norlux}
\format{
A vector of character strings.
}
\usage{
norlux
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for the NHXRGB0905005 light emitting diode (LEDs) arrays from Norlux (USA). 
Data are normalized to one at the wavelength of maximum emission.
}
\note{
Norlux is now part of Thomas Research Products.
}
\examples{
norlux

}
\references{
\url{http://www.trpssl.com/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{hewlett_packard}},
\code{\link{huey_jann}},
\code{\link{led_engin}},
\code{\link{leds_global}},
\code{\link{lumitronix}},
\code{\link{marktech}},
\code{\link{nichia}},
\code{\link{osram}},
\code{\link{quantum_devices}},
\code{\link{roithner_laser}},
\code{\link{seti}},
\code{\link{tao_yuan}},
\code{\link{unknown}}
}
\concept{manufacturers}
\keyword{datasets}
