% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osram.r
\docType{data}
\name{osram}
\alias{osram}
\title{Spectral data for LEDs array supplied by Osram}
\format{
A vector of character strings.
}
\usage{
osram
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for light emitting diodes (LEDs) from Osram. 
Data are normalized to one at the wavelength of maximum emission.
}
\note{
Current trade name is Osram Opto Semiconductors
}
\examples{
osram

}
\references{
\url{http://www.osram-os.com/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{hewlett_packard}},
\code{\link{huey_jann}},
\code{\link{led_engin}},
\code{\link{leds_global}},
\code{\link{lumitronix}},
\code{\link{marktech}},
\code{\link{nichia}},
\code{\link{norlux}},
\code{\link{quantum_devices}},
\code{\link{roithner_laser}},
\code{\link{seti}},
\code{\link{tao_yuan}},
\code{\link{unknown}}
}
\concept{manufacturers}
\keyword{datasets}
