% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photobiologyLEDs.r
\docType{package}
\name{photobiologyLEDs-package}
\alias{photobiologyLEDs}
\alias{photobiologyLEDs-package}
\title{photobiologyLEDs: Spectral Data for Light-Emitting-Diodes}
\description{
Spectral emission data for some frequently used light emitting
    diodes available as electronic components. Part of the 'r4photobiology' 
    suite, Aphalo P. J. (2015) <doi:10.19232/uv4pb.2015.1.14>.
}
\details{
Data for emission spectra of different types of LEDs.

The package contains one collection of spectra for different LEDs available
as electronic components requiring separate electronic driver modules or
circuits for their use. Data for LED bulbs are included in package 
\code{\link[photobiologyLamps]{photobiologyLamps-package}}. 

In addition to the spectra the packages provides character vectors of names
to be used as indexes to subset groups of spectra. In many cases spectral
data are normalized to spectral energy irradiance equal to one at the
wavelength of maximum spectral energy irradiance (strongest emission peak).

All LEDs have been measured at room temperature mounted on passive heatsinks
and driven at or below their maximum current rating. High precision power
supplies were used to drive them.
}
\section{Warning!}{

None of the spectral data included in this package are based on supplier's
specifications and are only for information. The exact emission spectrum
depends to some extent on testing conditions, but more importantly among
individual LED dies. Spectral specifications are usually given by typical and
boundary values. Furthermore, most manufacturers classify LEDs of a given
type into "bins" with slightly different optical and electrical
characteristics. In other words, the data provided here are not a substitute
for actual measurements of radiation emission and spectrum of the LEDs
actually used in a given piece of scientific research. For less demanding
situations the data are in most cases reliable enough but perfect agreement
with measurements on other LEDs of the same exact type should not be
expected.
}

\examples{
library(photobiology)
library(photobiologyWavebands)
library(ggspectra)

names(leds.mspct)

q_ratio(leds.mspct$white, Blue(), Red())

peaks(leds.mspct$white, span = 101)

autoplot(leds.mspct$white, annotations = c("+", "title:what"))

q_ratio(leds.mspct$Q36_4000K, Blue(), Red())

\dontrun{
autoplot(leds.mspct$Q36_4000K)
}

q_ratio(leds.mspct$NS6L183AT_H1, Blue(), Red())

\dontrun{
autoplot(leds.mspct$NS6L183AT_H1)
}

\dontrun{
autoplot(leds.mspct$NS6L183AT_H1, unit.out = "photon")
}

\dontrun{
autoplot(leds.mspct$NS6L183AT_H1, 
     range = VIS(), 
     w.band = VIS_bands(),
     span = 101)
}

}
\references{
Aphalo, Pedro J. (2015) The r4photobiology suite. UV4Plants Bulletin, 2015:1,
21-29. \url{https://doi.org/10.19232/uv4pb.2015.1.14}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.r4photobiology.info/photobiologyLEDs}
  \item \url{https://bitbucket.org/aphalo/photobiologyleds}
  \item Report bugs at \url{https://bitbucket.org/aphalo/photobiologyleds/issues}
}

}
\author{
\strong{Maintainer}: Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi} (\href{https://orcid.org/0000-0003-3385-972X}{ORCID})

Other contributors:
\itemize{
  \item Shafiuddin Ahmed [contributor]
}

}
