% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum-devices.r
\docType{data}
\name{quantum_devices}
\alias{quantum_devices}
\title{Spectral data for LEDs array supplied by Quantum Devices}
\format{
A vector of character strings.
}
\usage{
quantum_devices
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for light emitting diodes (LEDs) from Quantum Devices (USA). Data are
normalized to one at the wavelength of maximum emission.
}
\note{
Quantum Devices produces both individual LEDs and luminaires.
}
\examples{
quantum_devices

}
\references{
\url{http://www.quantumdev.com/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{hewlett_packard}},
\code{\link{huey_jann}},
\code{\link{led_engin}},
\code{\link{leds_global}},
\code{\link{lumitronix}},
\code{\link{marktech}},
\code{\link{nichia}},
\code{\link{norlux}},
\code{\link{osram}},
\code{\link{roithner_laser}},
\code{\link{seti}},
\code{\link{tao_yuan}},
\code{\link{unknown}}
}
\concept{manufacturers}
\keyword{datasets}
