% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seti.r
\docType{data}
\name{seti}
\alias{seti}
\title{Spectral data for LEDs array supplied by SETi}
\format{
A vector of character strings.
}
\usage{
seti
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance
for light emitting diodes (LEDs) arrays from SETi.
Data are normalized to one at the wavelength of maximum emission.
}
\note{
SETi (Sensor Electronic Technologies) is a supplier of high power ultraviolet
LEDs emitting in the UVC, UVB and UVA regions of the spectrum. Many of these
LEDs are also sold under different type denominations by Roithner
LaserTechnik.
}
\examples{
seti

}
\references{
<<<<<<< HEAD
<<<<<<< HEAD
\url{http://www.s-et.com/}
=======
\url{https://www.s-et.com/}
>>>>>>> minimal.fix.for.CRAN
=======
\url{https://www.s-et.com/}
>>>>>>> minimal.fix.for.CRAN
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{hewlett_packard}},
\code{\link{huey_jann}},
\code{\link{led_engin}},
\code{\link{leds_global}},
\code{\link{lumitronix}},
\code{\link{marktech}},
\code{\link{nichia}},
\code{\link{norlux}},
\code{\link{osram}},
\code{\link{quantum_devices}},
\code{\link{roithner_laser}},
\code{\link{tao_yuan}},
\code{\link{unknown}}
}
\concept{manufacturers}
\keyword{datasets}
