% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tao-yuan.r
\docType{data}
\name{tao_yuan}
\alias{tao_yuan}
\title{Spectral data for LEDs supplied by TaoYuan}
\format{
A vector of character strings.
}
\usage{
tao_yuan
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance
for different light emitting diodes (LEDs) from TaoYuan Electron (HK). Data
are normalized to one at the wavelength of maximum emission.
}
\note{
TaoYuan Electron (HK) is a supplier of LEDs and LED arrays.
}
\examples{
tao_yuan

}
\references{
\url{http://www.ledwv.com/en/}
}
\seealso{
\code{\link{leds.mspct}}

Other manufacturers: 
\code{\link{hewlett_packard}},
\code{\link{huey_jann}},
\code{\link{led_engin}},
\code{\link{leds_global}},
\code{\link{lumitronix}},
\code{\link{marktech}},
\code{\link{nichia}},
\code{\link{norlux}},
\code{\link{osram}},
\code{\link{quantum_devices}},
\code{\link{roithner_laser}},
\code{\link{seti}},
\code{\link{unknown}}
}
\concept{manufacturers}
\keyword{datasets}
