% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.r
\docType{data}
\name{unknown}
\alias{unknown}
\title{Spectral data for LEDs array of unknown manufacturer}
\format{
A vector of character strings.
}
\usage{
unknown
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance
for different "generic" light emitting diodes (LEDs) without type
specifications. Bought from shops like Class Ohlson or hobby targeted
electronic suppliers. Absolute values are not meaningful as the measuring
distances are variable, and in most cases unknown.
}
\examples{
unknown

}
\seealso{
Other manufacturers: 
\code{\link{hewlett_packard}},
\code{\link{huey_jann}},
\code{\link{led_engin}},
\code{\link{leds_global}},
\code{\link{lumitronix}},
\code{\link{marktech}},
\code{\link{nichia}},
\code{\link{norlux}},
\code{\link{osram}},
\code{\link{quantum_devices}},
\code{\link{roithner_laser}},
\code{\link{seti}},
\code{\link{tao_yuan}}
}
\concept{manufacturers}
\keyword{datasets}
