% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\docType{data}
\name{uv_leds}
\alias{uv_leds}
\alias{'leds}
\alias{by}
\alias{color'}
\alias{amber_leds}
\alias{blue_leds}
\alias{green_leds}
\alias{red_leds}
\alias{white_leds}
\alias{cyan_leds}
\alias{violet_leds}
\alias{multichannel_leds}
\title{Spectral data for LEDs of different colors}
\format{
A vector of character strings.

An object of class \code{character} of length 1.

An object of class \code{character} of length 5.

An object of class \code{character} of length 4.

An object of class \code{character} of length 11.

An object of class \code{character} of length 6.

An object of class \code{character} of length 2.

An object of class \code{character} of length 2.

An object of class \code{character} of length 1.
}
\usage{
uv_leds

amber_leds

blue_leds

green_leds

red_leds

white_leds

cyan_leds

violet_leds

multichannel_leds
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance 
for the light emitting diodes (LEDs) from various suppliers.
}
\examples{
uv_leds
violet_leds
blue_leds
cyan_leds
green_leds
amber_leds
red_leds
white_leds
multichannel_leds

# select LEDs emitting in the amber, yellow, orange region
leds.mspct[amber_leds]

}
\seealso{
\code{\link{leds.mspct}}
}
\keyword{datasets}
